+++ apps/api/internal/store/aim_trainer_store/db_get_user.go
@@ -8,7 +8,7 @@ import (
 
 type AimTrainerRow struct {
 	Username        string
-	Score           int64
+	Score           float64
 	Accuracy        float64
 	CompletedRounds int64
 	MiloAttacks     int64
